Database migrations

+++ partEZ/database/migrations/2016_01_27_004532_create_events_table.php
@@ -12,18 +12,21 @@ class CreateEventsTable extends Migration
      */
     public function up()
     {
-        Schema::create('events', function (Blueprint $table) {
-            $table->increments('eid')->unsigned();
-            $table->integer('uid')->unsigned();
-            $table->string('name');
-            $table->string('location');
-            $table->string('description');
-            $table->string('date');
-            $table->string('stime');
-            $table->string('etime');
-            $table->timestamps();
-            $table->foreign('uid')->references('uid')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('events'))
+        {
+            Schema::create('events', function (Blueprint $table) {
+                $table->increments('eid')->unsigned();
+                $table->integer('uid')->unsigned();
+                $table->string('name');
+                $table->string('location');
+                $table->string('description');
+                $table->string('date');
+                $table->string('stime');
+                $table->string('etime');
+                $table->timestamps();
+                $table->foreign('uid')->references('uid')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -33,6 +36,9 @@ class CreateEventsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('events');
+        if (Schema::hasTable('events'))
+        {
+            Schema::drop('events');
+        }
     }
 }
