+++ partEZ/database/migrations/2016_01_27_004619_create_invites_table.php
@@ -12,14 +12,17 @@ class CreateInvitesTable extends Migration
      */
     public function up()
     {
-        Schema::create('invites', function (Blueprint $table) {
-            $table->integer('eid')->unsigned();
-            $table->integer('uid')->unsigned();
-            $table->timestamps();
-            $table->primary('eid', 'uid');
-            $table->foreign('eid')->references('eid')->on('events')->onDelete('cascade');
-            $table->foreign('uid')->references('uid')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('invites'))
+        {
+            Schema::create('invites', function (Blueprint $table) {
+                $table->integer('eid')->unsigned();
+                $table->integer('uid')->unsigned();
+                $table->timestamps();
+                $table->primary('eid', 'uid');
+                $table->foreign('eid')->references('eid')->on('events')->onDelete('cascade');
+                $table->foreign('uid')->references('uid')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -29,6 +32,9 @@ class CreateInvitesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('invites');
+        if (Schema::hasTable('invites'))
+        {
+            Schema::drop('invites');
+        }
     }
 }
